Fixed test
